Create interface and refactor
